ARTEMIS-1467 clean up example
- Rename example project
- Leverage built-in 5.x Camel support instead of using WAR
- Clarify instructions
- Fix pom names/structure
