--- examples/features/sub-modules/artemis-jms-bridge/activemq-artemis-camel/src/main/resources/bridge.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:camel="http://camel.apache.org/schema/spring"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-                           http://camel.apache.org/schema/spring http://camel.apache.org/schema/spring/camel-spring.xsd">
-
-  <bean id="artemisConnectionFactory"
-        class="org.apache.activemq.ActiveMQConnectionFactory">
-    <property name="brokerURL" value="tcp://localhost:61616"/>
-    <property name="userName" value="admin"/>
-    <property name="password" value="password"/>
-  </bean>
-
-  <bean id="jmsConfig"
-        class="org.apache.camel.component.jms.JmsConfiguration">
-    <property name="connectionFactory" ref="artemisConnectionFactory"/>
-    <property name="concurrentConsumers" value="10"/>
-  </bean>
-
-  <bean id="activemq"
-        class="org.apache.activemq.camel.component.ActiveMQComponent">
-    <property name="configuration" ref="jmsConfig"/>
-  </bean>
-
-  <bean id="artemisConnectionFactory" class="org.apache.activemq.artemis.jms.client.ActiveMQConnectionFactory">
-    <constructor-arg name="url" value="tcp://localhost:61617"/>
-    <constructor-arg name="user" value="admin"/>
-    <constructor-arg name="password" value="password"/>
-  </bean>
-
-  <bean id="artemisConfig"
-        class="org.apache.camel.component.jms.JmsConfiguration">
-    <property name="connectionFactory" ref="artemisConnectionFactory"/>
-    <property name="concurrentConsumers" value="10"/>
-  </bean>
-
-  <bean id="artemis"
-        class="org.apache.camel.component.jms.JmsComponent">
-    <property name="configuration" ref="artemisConfig"/>
-  </bean>
-  
-
-  
-  <camelContext id="bridgeContext" trace="false" xmlns="http://camel.apache.org/schema/spring">
-    
-    <route id="bridge_TEST.FOO">
-       <from uri="activemq:queue:TEST.FOO"/>
-       <to uri="artemis:queue:TEST.FOO"/>
-    </route>
-    
-  </camelContext>
-
-</beans>
