--- examples/features/sub-modules/artemis-jms-bridge/pom.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>org.apache.artemis.examples</groupId>
-  <artifactId>artemis-jms-bridge-parent</artifactId>
-  <version>1.0.0-SNAPSHOT</version>
-
-  <packaging>pom</packaging>
-
-  <name>Active Messaging Bridge Examples</name>
-  <description>An example project showing different ways to use bridges with ActiveMQ and Artemis.</description>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-
-    <maven.compiler.source>8</maven.compiler.source>
-    <maven.compiler.target>8</maven.compiler.target>
-    
-    <maven-war-plugin.version>3.1.0</maven-war-plugin.version>
-
-    <camel.version>2.18.1</camel.version>
-    <spring.version>4.3.4.RELEASE</spring.version>
-    <slf4j-jboss-logging.version>1.1.0.Final</slf4j-jboss-logging.version>
-  </properties>
-
-  <modules>
-    <module>activemq-artemis-camel</module>
-    <module>artemis-jms-bridge</module>
-  </modules>
-
-  <repositories>
-    <repository>
-      <releases>
-        <enabled>true</enabled>
-        <updatePolicy>never</updatePolicy>
-      </releases>
-      <snapshots>
-        <enabled>false</enabled>
-      </snapshots>
-      <id>fuse-public-repository</id>
-      <name>FuseSource Community Release Repository</name>
-      <url>https://repo.fusesource.com/nexus/content/groups/public</url>
-    </repository>
-  </repositories>
-
-</project>
