+++ examples/features/sub-modules/inter-broker-bridge/artemis-jms-bridge/src/main/resources/bridge.xml
@@ -16,15 +16,14 @@ limitations under the License.
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-   <bean id="activemqConnectionFactory"
-         class="org.apache.activemq.ActiveMQConnectionFactory">
+   <bean id="5xConnectionFactory" class="org.apache.activemq.ActiveMQConnectionFactory">
       <property name="brokerURL" value="tcp://localhost:61616"/>
       <property name="userName" value="admin"/>
       <property name="password" value="password"/>
    </bean>
 
-   <bean id="activemqConnectionFactoryFactory" class="org.artemis.bridge.ConnectionFactoryFactoryBean">
-      <constructor-arg ref="activemqConnectionFactory"/>
+   <bean id="5xConnectionFactoryFactory" class="org.artemis.bridge.ConnectionFactoryFactoryBean">
+      <constructor-arg ref="5xConnectionFactory"/>
    </bean>
 
    <bean id="artemisConnectionFactory" class="org.apache.activemq.artemis.jms.client.ActiveMQConnectionFactory">
@@ -43,7 +42,7 @@ limitations under the License.
       <constructor-arg value="TEST.BAR"/>
    </bean>
 
-   <bean id="activemqDestinationFactory" class="org.artemis.bridge.DestinationFactoryBean">
+   <bean id="5xDestinationFactory" class="org.artemis.bridge.DestinationFactoryBean">
       <constructor-arg ref="activemqDestination"/>
    </bean>
 
@@ -54,9 +53,9 @@ limitations under the License.
    <bean id="artemis-jms-bridge" class="org.apache.activemq.artemis.jms.bridge.impl.JMSBridgeImpl" init-method="start" destroy-method="stop">
       <property name="bridgeName" value="artemis-jms-bridge"/>
       <property name="sourceConnectionFactoryFactory" ref="artemisConnectionFactoryFactory"/>
-      <property name="targetConnectionFactoryFactory" ref="activemqConnectionFactoryFactory"/>
+      <property name="targetConnectionFactoryFactory" ref="5xConnectionFactoryFactory"/>
       <property name="sourceDestinationFactory" ref="artemisDestinationFactory"/>
-      <property name="targetDestinationFactory" ref="activemqDestinationFactory"/>
+      <property name="targetDestinationFactory" ref="5xDestinationFactory"/>
       <property name="qualityOfServiceMode" value="AT_MOST_ONCE"/>
       <property name="failureRetryInterval" value="5000"/>
       <property name="failoverTimeout" value="5000"/>
@@ -64,6 +63,4 @@ limitations under the License.
       <property name="maxBatchSize" value="1"/>
       <property name="maxBatchTime" value="-1"/>
    </bean>
-
-
 </beans>
