+++ examples/features/sub-modules/inter-broker-bridge/artemis-jms-bridge/src/main/webapp/WEB-INF/web.xml
