+++ examples/features/sub-modules/pom.xml
@@ -50,7 +50,7 @@ under the License.
          <id>release</id>
          <modules>
             <module>artemis-ra-rar</module>
-	    <module>artemis-jms-bridge</module>
+	         <module>inter-broker-bridge</module>
          </modules>
       </profile>
    </profiles>
